pyproject.toml
[tool.poetry]
name = "hackathon"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
telebot = "^0.0.5"
matplotlib = "^3.9.1"
icalendar = "^5.0.13"
pillow = "^10.4.0"
python-telegram-bot = "^21.4"
requests = "^2.32.3"
google-generativeai = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
